+++ cards/moengage_cards_android/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add support for AGP `8.0.2` and above
 - Upgrade Kotlin Version to `1.7.10`
 - Support for `cards-core` version `1.6.0` and above
+- Fix: Android to Dart Method Channel Communication Breaking when `FirebaseMessaging.onBackgroundMessage` is used
 
 # 13-09-2023